api/requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.68.1
h11==0.12.0
importlib-metadata==4.8.1
pydantic==1.8.2
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
zipp==3.5.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
importlib-metadata==4.8.1
install==1.3.4
pydantic==1.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
zipp==3.5.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
importlib-metadata==4.8.1
install==1.3.4
pydantic==1.8.2
PyMySQL==1.0.2
python-dotenv==0.19.0
python-multipart==0.0.5
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
zipp==3.5.0
